add scf support
