+++ .gitignore
@@ -1,2 +1,6 @@
 
+xgdYqtb/*/
+
+xgdYqtb/*.so
+
 xgdYqtb/xgd_local.py
