+++ requirements.txt
@@ -1,3 +1,3 @@
-requests
-PyCryptodome
+cryptography
 lxml
+requests
